# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CodeVault API — AI software engineer that builds and pushes to GitHub repos
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface User {
  id: string;
  githubId: number;
  login: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  avatarUrl: string;
}

export interface Repo {
  id: number;
  name: string;
  fullName: string;
  /** @nullable */
  description?: string | null;
  private: boolean;
  defaultBranch: string;
  url: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface RepoCreate {
  /** @minLength 1 */
  name: string;
  description?: string;
  private?: boolean;
  autoInit?: boolean;
}

export type WorkspaceStatus = typeof WorkspaceStatus[keyof typeof WorkspaceStatus];


export const WorkspaceStatus = {
  ready: 'ready',
  cloning: 'cloning',
  error: 'error',
} as const;

export interface Workspace {
  id: string;
  repoFullName: string;
  branch: string;
  status: WorkspaceStatus;
  createdAt: string;
  /** @nullable */
  lastAccessedAt?: string | null;
}

export interface WorkspaceOpen {
  repoFullName: string;
  branch: string;
}

export interface WorkspaceStats {
  totalFiles: number;
  totalMessages: number;
  currentBranch: string;
  /** @nullable */
  lastCommitMessage?: string | null;
  /** @nullable */
  lastCommitDate?: string | null;
  uncommittedChanges?: number;
}

export type FileNodeType = typeof FileNodeType[keyof typeof FileNodeType];


export const FileNodeType = {
  file: 'file',
  dir: 'dir',
} as const;

export interface FileNode {
  path: string;
  name: string;
  type: FileNodeType;
  children?: FileNode[];
}

export interface FileContent {
  path: string;
  content: string;
  language: string;
}

export interface FileWrite {
  content: string;
}

export interface ReadFileInput {
  path: string;
}

export interface WriteFileInput {
  path: string;
  content: string;
}

export type FileDiffStatus = typeof FileDiffStatus[keyof typeof FileDiffStatus];


export const FileDiffStatus = {
  added: 'added',
  modified: 'modified',
  deleted: 'deleted',
  renamed: 'renamed',
} as const;

export interface FileDiff {
  path: string;
  status: FileDiffStatus;
  /** @nullable */
  diff?: string | null;
  additions: number;
  deletions: number;
  /** @nullable */
  oldPath?: string | null;
}

export interface Branch {
  name: string;
  current: boolean;
  /** @nullable */
  lastCommit?: string | null;
}

export interface BranchSwitch {
  branch: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export type FileChangeAction = typeof FileChangeAction[keyof typeof FileChangeAction];


export const FileChangeAction = {
  create: 'create',
  modify: 'modify',
  delete: 'delete',
} as const;

export interface FileChange {
  path: string;
  /** @nullable */
  content?: string | null;
  action: FileChangeAction;
}

export interface Message {
  id: string;
  workspaceId: string;
  role: MessageRole;
  content: string;
  fileChanges?: FileChange[];
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface AiResponse {
  message: Message;
  fileChanges: FileChange[];
}

export interface CommitInput {
  /** @nullable */
  message?: string | null;
}

export interface CommitResult {
  sha: string;
  message: string;
  branch: string;
  filesChanged: number;
}