# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeVault API — AI software engineer that builds and pushes to GitHub repos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List all open workspaces for the current user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: openWorkspace
      tags: [workspaces]
      summary: Clone and open a repository as a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceOpen"
      responses:
        "201":
          description: Opened workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by ID
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: closeWorkspace
      tags: [workspaces]
      summary: Close and clean up a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Workspace closed
  /workspaces/{workspaceId}/stats:
    get:
      operationId: getWorkspaceStats
      tags: [workspaces]
      summary: Get stats for a workspace (files, messages, last commit)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceStats"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      required: [id, repoFullName, branch, status, createdAt]
      properties:
        id:
          type: string
        repoFullName:
          type: string
        branch:
          type: string
        status:
          type: string
          enum: [ready, cloning, error]
        createdAt:
          type: string
        lastAccessedAt:
          type: ["string", "null"]
    WorkspaceOpen:
      type: object
      required: [repoFullName, branch]
      properties:
        repoFullName:
          type: string
        branch:
          type: string
    WorkspaceStats:
      type: object
      required: [totalFiles, totalMessages, currentBranch]
      properties:
        totalFiles:
          type: integer
        totalMessages:
          type: integer
        currentBranch:
          type: string
        lastCommitMessage:
          type: ["string", "null"]
        lastCommitDate:
          type: ["string", "null"]
        uncommittedChanges:
          type: integer
...